requirements/development.txt
# Inherit base requirements
-r base.txt

# Development tools
django-debug-toolbar==6.2.0
django-extensions==4.1
ipython==9.10.0
ipdb==0.13.13
werkzeug==3.1.6

# Testing
pytest==8.3.5
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.3
faker==40.8.0

# Code quality
black==25.3.1
flake8==7.1.2
isort==6.0.1
pre-commit==4.5.1

# Monitoring
django-silk==5.3.1